fix the incorrect Intel NVMe device path
1. No matter whether the NVMe device is from  Samsung or Intel, the NVMe
device path is /dev/libnvm* , this is defined by the BaM NVMe kernel module,
so for Intel NVMe devices we should not use /dev/libinvm*, which will lead to error.
2. the index of libnvm* should be "0,1,2,3,4..." instead of "0,1,4,9,2...", because if
there are only 3 NVMe devices for BaM, the third one will be failed to be open because
there's no /dev/libnvm4 in the host.

Signed-off-by: Liang Zhou <zhouliang.001@bytedance.com>